Fix repository to use Integer for primary key (Not long) and add show action

+++ src/main/java/hope/TodoRepository.java
@@ -7,6 +7,6 @@ import hope.Todo;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called todoRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface TodoRepository extends CrudRepository<Todo, Long> {
+public interface TodoRepository extends CrudRepository<Todo, Integer> {
 
 }
